backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
anthropic==0.34.2
openai==1.51.0
pinecone-client==5.0.1
supabase==2.9.0
upstash-redis==1.3.0
httpx==0.27.2
python-multipart==0.0.12
resend==2.4.0
stripe==10.10.0
pypdf==4.3.1
python-docx==1.1.2
python-jose[cryptography]==3.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
